+++ ibizpms-core/src/main/java/cn/ibizlab/pms/core/ibiz/service/impl/IbzCaseServiceImpl.java
@@ -113,7 +113,7 @@ public class IbzCaseServiceImpl extends ServiceImpl<IbzCaseMapper, IbzCase> impl
     }
     @Override
     @Transactional
-    public IbzCase get(BigInteger key) {
+    public IbzCase get(Long key) {
         IbzCase et = getById(key);
         if(et==null){
             et=new IbzCase();
@@ -168,22 +168,22 @@ public class IbzCaseServiceImpl extends ServiceImpl<IbzCaseMapper, IbzCase> impl
 
 
 	@Override
-    public List<IbzCase> selectByModule(BigInteger id) {
+    public List<IbzCase> selectByModule(Long id) {
         return baseMapper.selectByModule(id);
     }
 
     @Override
-    public void removeByModule(BigInteger id) {
+    public void removeByModule(Long id) {
         this.remove(new QueryWrapper<IbzCase>().eq("module",id));
     }
 
 	@Override
-    public List<IbzCase> selectByLib(BigInteger id) {
+    public List<IbzCase> selectByLib(Long id) {
         return baseMapper.selectByLib(id);
     }
 
     @Override
-    public void removeByLib(BigInteger id) {
+    public void removeByLib(Long id) {
         this.remove(new QueryWrapper<IbzCase>().eq("lib",id));
     }
