+++ ibizpms-core/src/main/java/cn/ibizlab/pms/core/ibiz/service/impl/IbzProjectMemberServiceImpl.java
@@ -79,19 +79,19 @@ public class IbzProjectMemberServiceImpl extends ServiceImpl<IbzProjectMemberMap
 
     @Override
     @Transactional
-    public boolean remove(BigInteger key) {
+    public boolean remove(Long key) {
         boolean result=removeById(key);
         return result ;
     }
 
     @Override
-    public void removeBatch(Collection<BigInteger> idList) {
+    public void removeBatch(Collection<Long> idList) {
         removeByIds(idList);
     }
 
     @Override
     @Transactional
-    public IbzProjectMember get(BigInteger key) {
+    public IbzProjectMember get(Long key) {
         IbzProjectMember et = getById(key);
         if(et==null){
             et=new IbzProjectMember();
